Refactor
- Move project folder to root
- Switch to IntelliJ
