--- Elevators v2/.settings/org.eclipse.core.resources.prefs
@@ -1,4 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/com/pecacheu/elevators/Conf.java=UTF-8
-encoding//src/com/pecacheu/elevators/Main.java=UTF-8
-encoding//src/com/pecacheu/elevators/config.yml=UTF-8
