--- Elevators v2/build.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project name="Elevators v2" default="makejar" basedir=".">
-    <target name="makejar" description="Create a jar for the project">
-        <mkdir dir="target"></mkdir>
-        <jar destfile="target/${ant.project.name}.jar">
-            <fileset dir="bin"></fileset>
-            <fileset file="plugin.yml"></fileset>
-        </jar>
-    </target>
-</project>
