--- Elevators v2/src/com/pecacheu/elevators/config.yml
@@ -1,49 +1,0 @@
-### This is the config file for Pecacheu's Elevator plugin.
-### You can adjust values here to globally change the behavior of elevators.
-
-debug: false # Debug mode. You'll need the 'elevators.reload' permission.
-title: '&1[&3Elevator&1]' # Header for ELEVATOR signs.
-call: '&1[&3Call&1]' # Header for CALL signs.
-error: '[&4???&r]' # Header for non-working signs.
-selStart: '&8> &5' # Start of selected floor line.
-selEnd: ' &8<' # End of selected floor line.
-noDoor: '&1[nodoor]' # NoDoor status indicator.
-
-msgGotoStart: '&eTraveling to &a' # Start of player message.
-msgGotoEnd: '&e.' # End of player message.
-msgCall: '&eCalling elevator.' # Player message for CALL signs.
-
-# Status symbols displayed on CALL signs:
-
-# Apparently, Java's UTF-8 support is buggy, so using
-# escapes rather than the actual characters is a good idea.
-noMove:   '&4\u29BF  \u29BF  \u29BF  \u29BF  \u29BF  \u29BF'
-mAtLevel: '&3\u29BF  \u29BF  \u29BF  \u29BF  \u29BF  \u29BF'
-atLevel:  '&2\u29BF  \u29BF  \u29BF  \u29BF  \u29BF  \u29BF'
-callUp:   '&3\u25B2  \u25B2  \u25B2  \u25B2  \u25B2  \u25B2'
-up:       '&4\u25B3  \u25B3  \u25B3  \u25B3  \u25B3  \u25B3'
-callDown: '&3\u25BC  \u25BC  \u25BC  \u25BC  \u25BC  \u25BC'
-down:     '&4\u25BD  \u25BD  \u25BD  \u25BD  \u25BD  \u25BD'
-
-floorMaxRadius: 8  # Maximum floor size, in each direction from sign.
-updateDelay: 150   # Delay between updates while moving. Increase to reduce lag.
-doorHoldTime: 4000 # Timeout before elevator doors are closed.
-saveInterval: 15   # Auto-save frequency in minutes.
-
-# List of blocks allowed as elevator floors:
-# Speeds are in blocks-per-second.
-
-blockList:
-  IRON_BLOCK: 8
-  GOLD_BLOCK: 10
-  EMERALD_BLOCK: 12
-  DIAMOND_BLOCK: 15
-  LEGACY_STAINED_GLASS: 5
-  GLASS: 4
-
-doorBlock: 'GLASS_PANE' # Material used for elevator block-doors.
-
-# Stored elevator locations:
-
-### WARNING: DO NOT TOUCH THIS SETTING ###
-elevators: {}
